feat(webpack): update config for dev
add dev command to enable files watch during development. makes development process faster and
easier.
