+++ webpack.prod.js
@@ -1,29 +1,23 @@
 const merge = require('webpack-merge');
+const BundleAnalyzer = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
+const Visualizer = require('webpack-visualizer-plugin');
 const path = require('path');
-const glob = require('glob');
 const common = require('./webpack.common.js');
-
-function getFiles(filePattern) {
-  const files = {
-    'index.js': path.join(__dirname, 'lib/index.js'),
-  };
-  glob.sync(filePattern).forEach((file) => {
-    files[file.replace('lib/', '../components/')] = path.join(__dirname, file);
-  });
-  return files;
-}
 
 module.exports = merge(common, {
   devtool: '',
   mode: 'production',
-  entry: getFiles('lib/!(index.js)**/index.js'),
-  output: {
-    path: path.resolve(__dirname, './dist'),
-    filename: '[name]',
-    libraryTarget: 'commonjs2',
-  },
   externals: [
     'react',
     'react-dom',
+  ],
+  plugins: [
+    new BundleAnalyzer({
+      analyzerMode: 'static',
+      reportFilename: path.resolve(__dirname, 'stats/bundle_analyzer_stats.html'),
+    }),
+    new Visualizer({
+      filename: '../stats/visualizer_stats.html',
+    }),
   ],
 });
